[hipblaslt] Fixing build issues for gfx_950_mx_rebase (#4465)
## Motivation

Solving various issues blocking hipblaslt from built

## Technical Details

Solving various issues blocking hipblaslt from built

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

+++ projects/hipblaslt/tensilelite/src/DataTypes.cpp
@@ -87,7 +87,7 @@ namespace rocisa
         return "Invalid";
     }
 
-    float GetElementSize(rocisa::DataType d)
+    size_t GetElementSize(rocisa::DataType d)
     {
         switch(d)
         {
